backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
google-generativeai==0.8.4
python-dotenv==1.0.1
python-multipart==0.0.20
